Redo test file naming and namespace conventions
