+++ src/ystdlib/error_handling/CMakeLists.txt
@@ -2,7 +2,8 @@ cpp_library(
     NAME error_handling
     NAMESPACE ystdlib
     TESTS_SOURCES
-        test-Defs.hpp
-        test-Defs.cpp
+        test-constants.hpp
+        test-types.hpp
+        test-types.cpp
         test-ErrorCode.cpp
 )
