+++ src/ystdlib/error_handling/test-types.cpp
@@ -1,10 +1,22 @@
-#include "test-Defs.hpp"
+#include "test-types.hpp"
 
 #include <algorithm>
-#include <array>
 #include <string>
 #include <string_view>
 #include <system_error>
+
+#include "test-constants.hpp"
+
+using ystdlib::error_handling::test::AlwaysSuccessErrorCategory;
+using ystdlib::error_handling::test::AlwaysSuccessErrorCodeEnum;
+using ystdlib::error_handling::test::BinaryErrorCategory;
+using ystdlib::error_handling::test::BinaryErrorCodeEnum;
+using ystdlib::error_handling::test::cAlwaysSuccessErrorCategoryName;
+using ystdlib::error_handling::test::cBinaryTestErrorCategoryName;
+using ystdlib::error_handling::test::cFailureConditions;
+using ystdlib::error_handling::test::cFailureErrorMsg;
+using ystdlib::error_handling::test::cSuccessErrorMsg;
+using ystdlib::error_handling::test::cUnrecognizedErrorCode;
 
 template <>
 auto AlwaysSuccessErrorCategory::name() const noexcept -> char const* {
